==> src/crowdstream/music/dataset-build/2_build_rekordbox_xml.py <==
# drag and drop the audio files samples (mp3) from spotify (step 1) into recordbox playlist/gallery window. 
# ensure they are analyzed. 
# in rekordbox go to FIle, Eport Collection in xlm format. That would save a file with audio file location
# and bpm. Ex. "collection.xml"
